+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/Mage.java
@@ -30,6 +30,7 @@ import com.elmakers.mine.bukkit.blocks.BlockList;
 import com.elmakers.mine.bukkit.blocks.UndoBatch;
 import com.elmakers.mine.bukkit.plugins.magic.wand.LostWand;
 import com.elmakers.mine.bukkit.plugins.magic.wand.Wand;
+import com.elmakers.mine.bukkit.plugins.magic.wand.WandLevel;
 import com.elmakers.mine.bukkit.utilities.InventoryUtils;
 import com.elmakers.mine.bukkit.utilities.UndoQueue;
 import com.elmakers.mine.bukkit.utilities.borrowed.ConfigurationNode;
@@ -319,6 +320,11 @@ public class Mage implements CostReducer
 		{
 			event.setCancelled(true);
 		}
+	}
+	
+	public boolean isProtected()
+	{
+		return activeWand != null && activeWand.getDamageReduction() > WandLevel.maxProtection;
 	}
 	
 	public void onPlayerDamage(EntityDamageEvent event)