+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  - Add "Sunny Day" spell, remove day and night spells.
  - The camera spell will now work on most mobs (not zombies or creepers or skeletons, though...)
  - Shrink and camera will now work on some blocks (cactus, tnt, chests, melons, logs and pumpkins)
+ - Gather, push, push, force and disarm will ignore targets with Protection X
 
 ## 2.9.0